Moving the file to be `.mjs
