+++ plugins/css-blank-pseudo/src/is-valid-replacement.mjs
